requirements.txt
streamlit
psycopg2-binary
plotly
pandas
sqlalchemy